requirements.txt
Flask==3.1.2
requests==2.32.5
geolib==1.0.7
python-dotenv==1.0.0
gunicorn==21.2.0